RN-664: Moved change pulling tests inside /pullChanges
